requirements.txt
python-telegram-bot==20.7
openai==1.3.5
PyPDF2==3.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.12.1
pydantic==2.5.0
asyncpg==0.29.0
uvloop==0.19.0
